Made generic test2D
* modified quad testing to be useful for triangle testing as well.
* added ERROR() logging primitive which outputs to std::cerr and also
  quits.
* added ElementType enum class
